requirements/base.txt
asgiref==3.8.1
attrs==22.2.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
coverage==7.2.3
debugpy==1.6.6
distlib==0.3.4
Django==5.1.6
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.3
django-phonenumber-field==7.0.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==9.8.3
filelock==3.4.2
flake8==4.0.1
identify==2.4.6
idna==3.3
iniconfig==2.0.0
mccabe==0.6.1
nodeenv==1.6.0
packaging==23.0
phonenumbers==8.13.7
Pillow==10.2.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==5.0.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
twilio==7.3.0
urllib3==1.26.7
virtualenv==20.13.0
drf-spectacular==0.28.0
django-cors-headers==4.3.1

# Encryption
cryptography==42.0.5

# AWS Storage
django-storages==1.14.2
boto3==1.34.41
